B42: make Core require-safe; Boot registers watchdog; context lazy-requires Core; hard requires for timed actions
