+++ 42/media/lua/client/AutoForester_Context.lua
@@ -1,23 +1,25 @@
-local function dbg(m) print("[AutoForester] "..tostring(m)) end
-dbg("Context file loaded")
-
-local function say(pi, txt) local p=getSpecificPlayer(pi or 0); if p and p.Say then p:Say(txt) end end
+local function say(pi, txt)
+  local p = getSpecificPlayer and getSpecificPlayer(pi or 0)
+  if p and p.Say then p:Say(txt) end
+end
 
 local function getSafeSquare(pi, wos)
-  local sq = getMouseSquare and getMouseSquare() or nil
-  if sq then return sq end
+  local ms = _G.getMouseSquare
+  local sq = (type(ms)=="function") and ms() or nil
+  if sq and sq.getX then return sq end
   if wos then
-    local first = wos.get and wos:get(0) or wos[1]
-    if first and first.getSquare then local s=first:getSquare(); if s then return s end end
+    local first = (wos.get and wos:get(0)) or wos[1]
+    if first and first.getSquare then
+      local s = first:getSquare(); if s then return s end
+    end
   end
-  local p=getSpecificPlayer and getSpecificPlayer(pi or 0)
-  if p and p.getSquare then return p:getSquare() end
-  return nil
+  local p = getSpecificPlayer and getSpecificPlayer(pi or 0)
+  return p and p:getSquare() or nil
 end
 
 local function core()
   local ok, mod = pcall(require, "AutoForester_Core")
-  if not ok or type(mod)~="table" then
+  if not ok or type(mod) ~= "table" then
     print("[AutoForester][ERROR] require Core failed: "..tostring(mod))
     return nil
   end
@@ -25,36 +27,25 @@ local function core()
 end
 
 local function addMenu(pi, context, wos, test)
-  -- heartbeat
   context:addOption("AutoForester: Debug (hook loaded)", nil, function() say(pi,"AF: hook OK") end)
   if test then return end
-
   local sq = getSafeSquare(pi, wos)
 
   context:addOption("Designate Wood Pile Here", sq, function(targetSq)
-    local C = core()
-    if not C then
-      local p = getSpecificPlayer and getSpecificPlayer(pi or 0)
-      if p and p.Say then p:Say("AutoForester core didn't load. Check console.") end
-      return
-    end
-    C.setStockpile(targetSq or getSafeSquare(pi, wos))
+    local c=core(); if not c then say(pi,"AutoForester core didn’t load. Check console."); return end
+    targetSq = targetSq or getSafeSquare(pi, wos)
+    if targetSq then c.setStockpile(targetSq); say(pi,"Wood pile set.") end
   end)
 
   context:addOption("Auto-Chop Nearby Trees", sq, function()
-    local C = core()
-    if not C then
-      local p = getSpecificPlayer and getSpecificPlayer(pi or 0)
-      if p and p.Say then p:Say("AutoForester core didn't load. Check console.") end
-      return
-    end
+    local c=core(); if not c then say(pi,"AutoForester core didn’t load. Check console."); return end
     local p = getSpecificPlayer(pi or 0); if not p then say(pi,"No player"); return end
-    C.startJob(p)
+    c.startJob(p)
   end)
 
-  local c=core()
+  local c = core()
   if c and c.hasStockpile() then
-    context:addOption("Clear Wood Pile Marker", nil, function() c.clearStockpile() end)
+    context:addOption("Clear Wood Pile Marker", nil, function() c.clearStockpile(); say(pi,"Wood pile cleared.") end)
   end
 end
 
@@ -67,8 +58,6 @@ local function register()
       if Events.OnFillWorldObjectContextMenu and Events.OnFillWorldObjectContextMenu.Add then
         Events.OnFillWorldObjectContextMenu.Add(addMenu)
         print("[AutoForester] Context hook registered (OnCreatePlayer)")
-      else
-        print("[AutoForester][WARN] Context event unavailable; skipping")
       end
     end)
   end
